# internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Elastic Maintainer API",
    "version": "v1alpha1",
    "description": "Initial web-first API contract. Endpoints marked with 501 responses are registered skeletons whose domain execution lands in later phases."
  },
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/targets/{targetId}/credentials": {
      "put": {
        "operationId": "putTargetCredentials",
        "parameters": [
          {
            "$ref": "#/components/parameters/TargetID"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/CSRFToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialPutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credential metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialStatusResponse"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/APIError"
          },
          "503": {
            "$ref": "#/components/responses/APIError"
          },
          "400": {
            "$ref": "#/components/responses/APIError"
          },
          "401": {
            "$ref": "#/components/responses/APIError"
          },
          "403": {
            "$ref": "#/components/responses/APIError"
          },
          "413": {
            "$ref": "#/components/responses/APIError"
          },
          "415": {
            "$ref": "#/components/responses/APIError"
          },
          "422": {
            "$ref": "#/components/responses/APIError"
          },
          "404": {
            "$ref": "#/components/responses/APIError"
          }
        }
      },
      "delete": {
        "operationId": "deleteTargetCredentials",
        "parameters": [
          {
            "$ref": "#/components/parameters/TargetID"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/CSRFToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmedMutationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credential metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialStatusResponse"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/APIError"
          },
          "503": {
            "$ref": "#/components/responses/APIError"
          },
          "400": {
            "$ref": "#/components/responses/APIError"
          },
          "401": {
            "$ref": "#/components/responses/APIError"
          },
          "403": {
            "$ref": "#/components/responses/APIError"
          },
          "413": {
            "$ref": "#/components/responses/APIError"
          },
          "415": {
            "$ref": "#/components/responses/APIError"
          },
          "404": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-elastic-maintainer-session"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OIDC JWT access token, bounded to 16 KiB and validated with RS256 or ES256 against the configured pinned issuer, client audience, JWKS transport, and claim-to-role mapping. Bearer tokens do not require a browser nonce. Duplicate Authorization headers or any simultaneous application-session cookie are rejected."
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$"
        }
      },
...
      "TargetID": {
        "name": "targetId",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/SafeID"
        }
      },
...
      "CSRFToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": false,
        "description": "Required for browser-session credential mutations; omitted for bearer automation.",
        "schema": {
          "type": "string",
          "minLength": 32,
          "maxLength": 128
        }
      }
    },
    "responses": {
      "APIError": {
        "description": "API error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
      "APIVersion": {
        "type": "string",
        "enum": [
          "elastic-maintainer/v1alpha1"
        ]
      },
      "SafeID": {
        "type": "string",
        "minLength": 1,
        "maxLength": 64,
        "pattern": "^[A-Za-z0-9_-]+$"
      },
...
      "ConfirmedMutationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "confirm"
        ],
        "properties": {
          "confirm": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        }
      },
      "CredentialPutRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "apiKey"
        ],
        "properties": {
          "apiKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 16384,
            "writeOnly": true,
            "format": "password",
            "pattern": "^[!-~]+$"
          },
          "caCertificatePem": {
            "type": "string",
            "maxLength": 262144,
            "writeOnly": true
          }
        }
      },
      "CredentialStatus": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "configured"
        ],
        "properties": {
          "configured": {
            "type": "boolean"
          },
          "secretResourceVersion": {
            "type": "string"
          },
          "rotatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "rotatedBy": {
            "type": "string"
          },
          "certificateSHA256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "certificateNotAfter": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CredentialStatusResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "apiVersion",
          "targetId",
          "credentialStatus"
        ],
        "properties": {
          "apiVersion": {
            "$ref": "#/components/schemas/APIVersion"
          },
          "targetId": {
            "$ref": "#/components/schemas/SafeID"
          },
          "credentialStatus": {
            "$ref": "#/components/schemas/CredentialStatus"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "apiVersion",
          "error"
        ],
        "properties": {
          "apiVersion": {
            "$ref": "#/components/schemas/APIVersion"
          },
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}